Add JavaScript methods
