+++ dotnet/Liberator.Driver/Helpers.cs
@@ -54,4 +54,76 @@ namespace Liberator.Driver
             return null;
         }
     }
+
+    /// <summary>
+    /// The measurements for elements
+    /// </summary>
+    public class ElementSize
+    {
+        /// <summary>
+        /// Height of the element
+        /// </summary>
+        public int Height { get; set; }
+
+        /// <summary>
+        /// Leftmost edge of the element
+        /// </summary>
+        public int Left { get; set; }
+
+        /// <summary>
+        /// Topmost edge of the element
+        /// </summary>
+        public int Top { get; set; }
+
+        /// <summary>
+        /// Width of the element
+        /// </summary>
+        public int Width { get; set; }
+    }
+
+    /// <summary>
+    /// Bounding client rectangle
+    /// </summary>
+    public class DomRectangle
+    {
+        /// <summary>
+        /// Leftmost edge
+        /// </summary>
+        public int Left { get; set; }
+
+        /// <summary>
+        /// Topmost edge
+        /// </summary>
+        public int Top { get; set; }
+
+        /// <summary>
+        /// Rightmost edge
+        /// </summary>
+        public int Right { get; set; }
+
+        /// <summary>
+        /// Bottom edge
+        /// </summary>
+        public int Bottom { get; set; }
+
+        /// <summary>
+        /// X Co-ordinate
+        /// </summary>
+        public int X { get; set; }
+
+        /// <summary>
+        /// Y Co-ordinate
+        /// </summary>
+        public int Y { get; set; }
+
+        /// <summary>
+        /// Width of the element
+        /// </summary>
+        public int Width { get; set; }
+
+        /// <summary>
+        /// Height of the element
+        /// </summary>
+        public int Height { get; set; }
+    }
 }
