+++ dotnet/Liberator.Driver/RatDriverActions.cs
@@ -1231,164 +1231,5 @@ namespace Liberator.Driver
         //}
 
         #endregion
-
-        #region Additional Methods
-
-        /// <summary>
-        /// Takes a screenshot of the active web page
-        /// </summary>
-        /// <param name="path">The path and name under which to save the image</param>
-        public void TakeScreenshot(string path)
-        {
-            try
-            {
-                Screenshot s = ((ITakesScreenshot)Driver).GetScreenshot();
-                s.SaveAsFile(path, ScreenshotImageFormat.Jpeg);
-                Console.Out.WriteLine("Screenshot taken and saved to {0}.", path);
-            }
-            catch (Exception ex)
-            {
-                if (Preferences.BaseSettings.DebugLevel == EnumConsoleDebugLevel.Human)
-                {
-                    Console.Out.WriteLine("Could not take a screenshot.");
-                }
-                HandleErrors(ex);
-            }
-        }
-
-        /// <summary>
-        /// Returns a JavaScript Executor to the end user for custom JavaScript
-        /// </summary>
-        /// <returns>A JavaScript Executor</returns>
-        public IJavaScriptExecutor Scripts()
-        {
-            try
-            {
-                IJavaScriptExecutor js = ((IJavaScriptExecutor)Driver);
-                return js;
-            }
-            catch (Exception ex)
-            {
-                if (Preferences.BaseSettings.DebugLevel == EnumConsoleDebugLevel.Human)
-                {
-                    Console.Out.WriteLine("Could not return the JavaScript Executor.");
-                }
-                HandleErrors(ex);
-                return null;
-            }
-        }
-
-        /// <summary>
-        /// Executes a JavaScript statement passed as a string
-        /// </summary>
-        /// <param name="script">The JavaScript to be executed</param>
-        public void ExecuteJavaScript(string script)
-        {
-            try
-            {
-                IJavaScriptExecutor js = Scripts();
-                js.ExecuteScript(script);
-                Console.Out.WriteLine("Executed the sctipt '{0}'.", script);
-            }
-            catch (Exception ex)
-            {
-                if (Preferences.BaseSettings.DebugLevel == EnumConsoleDebugLevel.Human)
-                {
-                    Console.Out.WriteLine("Could not execute the passed JavaScript.");
-                }
-                HandleErrors(ex);
-            }
-        }
-
-        /// <summary>
-        /// Executes a JavaScript statement passed as a string
-        /// </summary>
-        /// <param name="script">The JavaScript to be executed</param>
-        /// <param name="parameters">The parameters for the passed JavaScript</param>
-        public void ExecuteJavaScript(string script, params object[] parameters)
-        {
-            try
-            {
-                IJavaScriptExecutor js = ((IJavaScriptExecutor)Driver);
-                js.ExecuteScript(script, parameters);
-                Console.Out.WriteLine("Executed the sctipt '{0}'.", script);
-            }
-            catch (Exception ex)
-            {
-                if (Preferences.BaseSettings.DebugLevel == EnumConsoleDebugLevel.Human)
-                {
-                    Console.Out.WriteLine("Could not execute the passed JavaScript with the specified parameters.");
-                }
-                HandleErrors(ex);
-            }
-        }
-
-        /// <summary>
-        /// Executes a JavaScript statement passed as a string
-        /// </summary>
-        /// <param name="script">The JavaScript to be executed</param>
-        public void ExecuteAsyncJavaScript(string script)
-        {
-            try
-            {
-                IJavaScriptExecutor js = ((IJavaScriptExecutor)Driver);
-                js.ExecuteAsyncScript(script);
-                Console.Out.WriteLine("Executed the sctipt '{0}'.", script);
-            }
-            catch (Exception ex)
-            {
-                if (Preferences.BaseSettings.DebugLevel == EnumConsoleDebugLevel.Human)
-                {
-                    Console.Out.WriteLine("Could not execute the passed asynchronous JavaScript.");
-                }
-                HandleErrors(ex);
-            }
-        }
-
-        /// <summary>
-        /// Executes a JavaScript statement passed as a string
-        /// </summary>
-        /// <param name="script">The JavaScript to be executed</param>
-        /// <param name="parameters">The parameters for the passed JavaScript</param>
-        public void ExecuteAsyncJavaScript(string script, params object[] parameters)
-        {
-            try
-            {
-                IJavaScriptExecutor js = ((IJavaScriptExecutor)Driver);
-                js.ExecuteAsyncScript(script, parameters);
-                Console.Out.WriteLine("Executed the sctipt '{0}'.", script);
-            }
-            catch (Exception ex)
-            {
-                if (Preferences.BaseSettings.DebugLevel == EnumConsoleDebugLevel.Human)
-                {
-                    Console.Out.WriteLine("Could not execute the passed asynchronous JavaScript.");
-                }
-                HandleErrors(ex);
-            }
-        }
-
-        /// <summary>
-        /// Scrolls to an element using JavaScript
-        /// </summary>
-        /// <param name="webElement"></param>
-        public void ScrollToElement(IWebElement webElement)
-        {
-            try
-            {
-                IJavaScriptExecutor js = ((IJavaScriptExecutor)Driver);
-                js.ExecuteScript("arguments[0].scrollIntoView(true);", webElement);
-            }
-            catch (Exception ex)
-            {
-                if (Preferences.BaseSettings.DebugLevel == EnumConsoleDebugLevel.Human)
-                {
-                    Console.Out.WriteLine("Could not execute the passed asynchronous JavaScript.");
-                }
-                HandleErrors(ex);
-            }
-        }
-
-        #endregion
     }
 }
